Package Internals

Highlights Module

Highlights.AbstractLexer — Type.

abstract AbstractLexer

Represents a source code lexer used to tokenise text.

source

Highlights.AbstractTheme — Type.

abstract AbstractTheme

Represents a colour scheme used to highlight tokenised source code.

source

Highlights.definition — Function.

This function is extended by each new subtype of AbstractLexer and AbstractTheme to provide the data nessesary to compile lexer functions and colour schemes.

Examples

abstract CustomLexer <: AbstractLexer

Highlights.definition(::Type{CustomLexer}) = Dict(
    :name => "Custom",
    :tokens => Dict(
        :root => [
            # Rules go here...
        ],
    ),
)

source

Highlights.lexer — Method.

lexer(name)

Return the AbstractLexer associated with the lexer named name. name must be a string. Internally this checks the :aliases field in the definition of a lexer to see whether it is a match.

Warning

This method is not type stable.

When no lexer matches the given name an ArgumentError is thrown.

source

Themes Submodule

Highlights.Themes.Style — Type.

immutable Style

An internal type used to track colour scheme definition information such as foreground and background colours as well as bold, italic, and underlining.

source

Highlights.Themes.Theme — Type.

immutable Theme

Represents a "compiled" colour scheme as constructed by build_theme. Not a user-facing type.

source

Highlights.Themes.build_theme — Method.

Build a colour scheme Theme object based on a user-defined theme and lexer.

source

Highlights.Themes.expand_html — Method.

expand_html(s)

Expands an HTML colour code s from 3 digits to 6.

source

Highlights.Themes.has_bg — Method.

has_bg(s)

Does the Style s have a background colour set?

source

Highlights.Themes.has_fg — Method.

has_fg(s)

Does the Style s have a foreground colour set?

source

Highlights.Themes.@S_str — Macro.

@S_str(spec)

Construct a Style object.

source

Lexers Submodule

Highlights.Lexers.words — Method.

words(vec; prefix, suffix)

Build a regular expression from a vector of strings. Useful for keyword lists.

The keyword arguments prefix and suffix can be used to add patterns before and after the main match group.

Examples

julia> import Highlights.Lexers: words

julia> words(["if", "else"])
r"(if|else)"

source

Highlights.Lexers.@raw_str — Macro.

@raw_str(str)

A utility string macro to avoid having to escape strings used to build a regular expression.

source