Public Interface

Highlights Module

Highlights — Module.

Highlights.jl is a Julia package for source code highlighting. It provides a regular expression based mechanism for creating lexers in a similar way to Pygments.

The following names are exported from the root module, Highlights, and are available for external use. Note that unexported names are considered unstable and subject to change.

Lexers

Themes

Tokens

highlight

lexer

stylesheet

source

Highlights.highlight — Function.

Highlight source code using a specific lexer, mimetype and theme.

highlight(io, mime, src, lexer, theme)
highlight(io, mime, src, lexer)

src is tokenised using the provided lexer, then colourised using theme, and finally output to io in the given format mime. theme defaults to Themes.DefaultTheme theme.

mime can be either MIME("text/html") or MIME("text/latex").

Examples

julia> using Highlights

julia> highlight(STDOUT, MIME("text/html"), "2x", Lexers.JuliaLexer)
<pre class='hljl'>
<span class='hljl-NUMBER_INTEGER'>2</span><span class='hljl-NAME'>x</span>
</pre>

julia> highlight(STDOUT, MIME("text/latex"), "'x'", Lexers.JuliaLexer, Themes.VimTheme)
\begin{lstlisting}
(*@\HLJLSTRINGCHAR{{\textquotesingle}x{\textquotesingle}}@*)
\end{lstlisting}

source

Highlights.lexer — Method.

lexer(name)

Return the AbstractLexer associated with the lexer named name. name must be a string. Internally this checks the :aliases field in each lexer definition to see whether it is a match.

Warning

This method is not type stable.

When no lexer matches the given name an ArgumentError is thrown.

source

Highlights.stylesheet — Function.

Generate a "stylesheet" for the given theme.

stylesheet(io, mime, theme)
stylesheet(io, mime)

Prints out the style information needed to colourise source code in the given theme. Note that theme defaults to Themes.DefaultTheme. Output is printed to io in the format mime. mine can be one of

MIME("text/html")

MIME("text/css")

MIME("text/latex")

Examples

julia> using Highlights

julia> buf = IOBuffer();

julia> stylesheet(buf, MIME("text/css"), Themes.EmacsTheme)

julia> split(takebuf_string(buf), '\n')[1] # Too much output to show everything.
"pre.hljl {"

source

Themes Submodule

Highlights.Themes — Module.

A submodule that provides a selection of themes that can be used to colourise source code.

AbstractTheme

DefaultTheme

EmacsTheme

GitHubTheme

MonokaiTheme

PygmentsTheme

@S_str

TangoTheme

TracTheme

VimTheme

VisualStudioTheme

XcodeTheme

@theme

source

Highlights.Themes.DefaultTheme — Type.

The default colour scheme with colours based on the Julia logo.

source

Highlights.Themes.EmacsTheme — Type.

A theme based on the Emacs colour scheme.

source

Highlights.Themes.GitHubTheme — Type.

A GitHub inspired colour scheme.

source

Highlights.Themes.MonokaiTheme — Type.

A colour scheme similar to the Monokai theme.

source

Highlights.Themes.PygmentsTheme — Type.

Based on the default colour scheme used by the Pygments highlighter.

source

Highlights.Themes.TangoTheme — Type.

A Tango-inspired colour scheme.

source

Highlights.Themes.TracTheme — Type.

Based on the default trac highlighter.

source

Highlights.Themes.VimTheme — Type.

A Vim 7.0 based colour scheme.

source

Highlights.Themes.VisualStudioTheme — Type.

A theme based on the default Visual Studio colours.

source

Highlights.Themes.XcodeTheme — Type.

A theme based on the default Xcode colour scheme.

source

Highlights.Themes.@S_str — Macro.

@S_str(spec)

Construct a Style object.

source

Highlights.Themes.@theme — Macro.

@theme(T, dict)

Declare the theme definition for theme T based on dict. dict must be a Dict and T must be a subtype of AbstractTheme.

Examples

julia> using Highlights.Themes

julia> abstract CustomTheme <: AbstractTheme

julia> @theme CustomTheme Dict(
           :name => "Custom",
           :tokens => Dict(
               # ...
           )
       );

source

Lexers Submodule

Highlights.Lexers — Module.

An exported submodule that provides a selection of lexer definitions for tokenising source code. The following names are exported from the module:

AbstractLexer

FortranLexer

JuliaConsoleLexer

JuliaLexer

MatlabLexer

RLexer

TOMLLexer

@lexer

source

Highlights.Lexers.FortranLexer — Type.

A FORTRAN 90 source code lexer.

source

Highlights.Lexers.JuliaConsoleLexer — Type.

A lexer for Julia REPL sessions.

source

Highlights.Lexers.JuliaLexer — Type.

A lexer for Julia source code.

source

Highlights.Lexers.MatlabLexer — Type.

A lexer for MATLAB source code.

source

Highlights.Lexers.RLexer — Type.

A lexer for the R language.

source

Highlights.Lexers.TOMLLexer — Type.

TOML (Tom's Obvious, Minimal Language) lexer.

source

Highlights.Lexers.@lexer — Macro.

@lexer(T, dict)

Declare the lexer definition for a custom lexer T using a Dict object dict. dict must either be a Dict literal or an expression that returns a Dict – such as a let block.

Examples

julia> using Highlights.Lexers

julia> abstract CustomLexer <: AbstractLexer

julia> @lexer CustomLexer Dict(
           :name => "Custom",
           :tokens => Dict(
               # ...
           )
       );

source

Tokens Subtype

Highlights.Tokens — Module.

This submodule provides a collection of token names for use in lexer and theme definitions. Exports the following list of tokens:

TEXT

WHITESPACE

ESCAPE

ERROR

OTHER

KEYWORD

KEYWORD_CONSTANT

KEYWORD_DECLARATION

KEYWORD_NAMESPACE

KEYWORD_PSEUDO

KEYWORD_RESERVED

KEYWORD_TYPE

NAME

NAME_ATTRIBUTE

NAME_BUILTIN

NAME_BUILTIN_PSEUDO

NAME_CLASS

NAME_CONSTANT

NAME_DECORATOR

NAME_ENTITY

NAME_EXCEPTION

NAME_FUNCTION

NAME_FUNCTION_MAGIC

NAME_PROPERTY

NAME_LABEL

NAME_NAMESPACE

NAME_OTHER

NAME_TAG

NAME_VARIABLE

NAME_VARIABLE_CLASS

NAME_VARIABLE_GLOBAL

NAME_VARIABLE_INSTANCE

NAME_VARIABLE_MAGIC

LITERAL

LITERAL_DATE

STRING

STRING_AFFIX

STRING_BACKTICK

STRING_CHAR

STRING_DELIMITER

STRING_DOC

STRING_DOUBLE

STRING_ESCAPE

STRING_HEREDOC

STRING_INTERPOL

STRING_OTHER

STRING_REGEX

STRING_SINGLE

STRING_SYMBOL

NUMBER

NUMBER_BIN

NUMBER_FLOAT

NUMBER_HEX

NUMBER_INTEGER

NUMBER_INTEGER_LONG

NUMBER_OCT

OPERATOR

OPERATOR_WORD

PUNCTUATION

COMMENT

COMMENT_HASHBANG

COMMENT_MULTILINE

COMMENT_PREPROC

COMMENT_PREPROCFILE

COMMENT_SINGLE

COMMENT_SPECIAL

GENERIC

GENERIC_DELETED

GENERIC_EMPH

GENERIC_ERROR

GENERIC_HEADING

GENERIC_INSERTED

GENERIC_OUTPUT

GENERIC_PROMPT

GENERIC_STRONG

GENERIC_SUBHEADING

GENERIC_TRACEBACK

source