Public Interface

Highlights Module

Highlights — Module.

Highlights.jl is a Julia package for source code highlighting. It provides a regular expression based mechanism for creating lexers in a similar way to Pygments.

The following names are exported from the root module, Highlights, and are available for external use. Note that unexported names are considered unstable and subject to change.

Lexers

Themes

highlight

stylesheet

source

Highlights.highlight — Function.

Highlight source code using a specific lexer, mimetype and theme.

highlight(io, mime, src, lexer, theme)
highlight(io, mime, src, lexer)

src is tokenised using the provided lexer, then colourised using theme, and finally output to io in the given format mime. theme defaults to Themes.DefaultTheme theme.

mime can be either of MIME("text/html") or MIME("text/latex").

Examples

julia> using Highlights

julia> highlight(STDOUT, MIME("text/html"), "2x", Lexers.JuliaLexer)
<pre class='hljl'>
<span class='hljl-NUMBER'>2</span><span class='hljl-NAME'>x</span>
</pre>

julia> highlight(STDOUT, MIME("text/latex"), "'x'", Lexers.JuliaLexer, Themes.VimTheme)
\begin{lstlisting}
(*@\HLJLSTRING{{\textquotesingle}x{\textquotesingle}}@*)
\end{lstlisting}

source

Highlights.stylesheet — Function.

Generate a "stylesheet" for the given theme.

stylesheet(io, mime, theme)
stylesheet(io, mime)

Prints out the style information needed to colourise source code in the given theme. Note that theme defaults to Themes.DefaultTheme. Output is printed to io in the format mime. mine can be one of

MIME("text/html")

MIME("text/css")

MIME("text/latex")

Examples

julia> using Highlights

julia> buf = IOBuffer();

julia> stylesheet(buf, MIME("text/css"), Themes.EmacsTheme)

julia> split(takebuf_string(buf), '\n')[1] # Too much output to show everything.
"pre.hljl {"

source

Themes Submodule

Highlights.Themes — Module.

A submodule that provides a selection of themes that can be used to colourise source code.

DefaultTheme

EmacsTheme

GitHubTheme

MonokaiTheme

PygmentsTheme

TangoTheme

TracTheme

VimTheme

VisualStudioTheme

XcodeTheme

source

Highlights.Themes.DefaultTheme — Type.

The default colour scheme with colours based on the Julia logo.

source

Highlights.Themes.EmacsTheme — Type.

A theme based on the Emacs colour scheme.

source

Highlights.Themes.GitHubTheme — Type.

A GitHub inspired colour scheme.

source

Highlights.Themes.MonokaiTheme — Type.

A colour scheme similar to the Monokai theme.

source

Highlights.Themes.PygmentsTheme — Type.

Based on the default colour scheme used by the Pygments highlighter.

source

Highlights.Themes.TangoTheme — Type.

A Tango-inspired colour scheme.

source

Highlights.Themes.TracTheme — Type.

Based on the default trac highlighter.

source

Highlights.Themes.VimTheme — Type.

A Vim 7.0 based colour scheme.

source

Highlights.Themes.VisualStudioTheme — Type.

A theme based on the default Visual Studio colours.

source

Highlights.Themes.XcodeTheme — Type.

A theme based on the default Xcode colour scheme.

source

Lexers Submodule

Highlights.Lexers — Module.

An exported submodule that provides a selection of lexer definitions for tokenising source code. The following lexers are exported from the module:

FortranLexer

JuliaConsoleLexer

JuliaLexer

MatlabLexer

TOMLLexer

source

Highlights.Lexers.FortranLexer — Type.

A FORTRAN 90 source code lexer.

source

Highlights.Lexers.JuliaConsoleLexer — Type.

A lexer for Julia REPL sessions.

source

Highlights.Lexers.JuliaLexer — Type.

A lexer for Julia source code.

source

Highlights.Lexers.MatlabLexer — Type.

A lexer for MATLAB source code.

source

Highlights.Lexers.TOMLLexer — Type.

TOML (Tom's Obvious, Minimal Language) lexer.

source

Tokens Subtype

Highlights.Tokens — Module.

This submodule provides a collection of token names for use in lexer and theme definitions. Additional tokens can be defined using the @tokengroup macro.

source

Highlights.Tokens.@tokengroup — Macro.

@tokengroup(name, tokens)

Define a new group of tokens with the given group name and individual names tokens where tokens is an array literal containing valid identifiers.

Examples

using Highlights.Tokens

@tokengroup CustomGroup [
    FOO,
    BAR,
    BAZ,
]

source